fix apartment

+++ src/main/java/lab/space/my_house_24/entity/Apartment.java
@@ -24,7 +24,7 @@ public class Apartment {
     private Integer number;
 
     @Column(nullable = false)
-    private Integer area;
+    private Double area;
 
     @ManyToOne
     @JoinColumn(name = "house_id")
