+++ src/main/java/lab/space/my_house_24/model/apartment/ApartmentAddRequest.java
@@ -13,7 +13,7 @@ public record ApartmentAddRequest(
 
 
         @NotNull(message = "{not.blank.message}")
-        Integer area,
+        Double area,
 
 
         @NotNull(message = "{not.blank.message}")
